+++ Auditive/Assets.xcassets/AppIcon.appiconset/Contents.json
