+++ Auditive/Assets.xcassets/Contents.json
